Moving request class to storage module.
