+++ graylog-storage-elasticsearch6/src/main/java/org/graylog/storage/elasticsearch6/cluster/GetAllocationDiskSettings.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.indexer.cluster.jest;
+package org.graylog.storage.elasticsearch6.cluster;
 
 import io.searchbox.cluster.GetSettings;
